Fix erroneously failing field tests
This commit fixes two issues with the field test functions in the test core.
Issue one arose when creating the threadErrs array, but not initializing it to
0. This causes undefined behavior that can be interpreted as non-zero return
codes from the tests when the tests are in fact passing. After setting all
values to 0 before the tests run, the field tests now correctly show that they
are passing.

The second issue comes from the use of mpas_threading_get_thread_num to get the
thread number that the tests are running on, but only running the tests when
threadNum == 0. If one of the tests does actually fail, threadErrs( threadNum )
is updated to 1. Since fortran has 1 based arrays, this led to out of bounds
errors in the event that a test actually did fail. This was fixed by setting
threadNum to 1 after chacking that threadNum == 0, meaning that we still update
the first element in threadErrs if there is a failed test. If these tests are
updated to run on more than one thread, threadNum should be updated to be
mpas_threading_get_thread_num() + 1, and threadNum references should be based on
that.